Extend Events to support tokens for listeners.
The `Event` and `EventListener` types are parameterized by a new type
`T`, which represents the type of the token that is passed with each
notifitication. This type is `()` by default, unless specified otherwise
by users.

This allows users to pass tokens from the notifier to the listener with
new `notify_with_token` APIs. The `EventListener` Future now resolves
to its token type, instead of always resolving to `()`.

An example use case is a `Mutex` which, like the parking_lot Mutex, can
pass the lock directly on to the next waiting listener, without ever
actually unlocking the Mutex's state.

This change was done backwards compatibly, so APIs which previous relied
on the absence of a token type (like the existing notify APIs, and some
of the wait APIs) are only provided by events which have no token,
whereas new APIs with tokens are added additionally.

Events with tokens can only notify one event at a type. It would be
interesting to extend this to support passing multiple tokens to notify
multiple listeners at once, with tokens.