Remove tag support from no_std
For tagging to work under no-std mode, the tags would need to be stored
in the backup queue. For notifications for std::usize::MAX listeners,
this is not possible. Therefore the only strategy left to use that doesn't
massively constrain the use of tagging is to remove tags from no_std
mode.

Closes #73

Signed-off-by: John Nunley <dev@notgull.net>