chore: Revert back to v2.5.3
I'm man enough to know when I've been wrong.

The intrusive-list-based list implementation is riddled with bugs, fails
MIRI, and overall increases the complexity of this crate. I've invested
a lot of time and energy into improving it; effort which I feel has been
wasted. So I figure I'll cut my losses and revert.

This PR undoes every commit made between now and the release of
event-listener version 2.5.3. The next commits will be replaying newer
features in order to avoid a breaking change.

Signed-off-by: John Nunley <dev@notgullnet>